+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.0.32] - 2023-02-15
+
+### Added
+
+- New Mod: `Tunnel Man`
+
 ## [0.0.31] - 2023-02-15
 
 ### Added